% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_expectation.R
\name{AdaptiveExpectation_calcExpectedPrice}
\alias{AdaptiveExpectation_calcExpectedPrice}
\title{AdaptiveExpectation_calcExpectedPrice}
\usage{
AdaptiveExpectation_calcExpectedPrice(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected price for}

\item{aPeriod}{Current model period}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
AdaptiveExpectation_calcExpectedPrice
}
\details{
Calculate the expected price for a LandLeaf using
         an adaptive approach -- use linear combination of previous expectation and
         new information.
}
\author{
KVC November 2017
}
