% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_expectation.R
\name{AdaptiveExpectation_calcExpectedYield}
\alias{AdaptiveExpectation_calcExpectedYield}
\title{AdaptiveExpectation_calcExpectedYield}
\usage{
AdaptiveExpectation_calcExpectedYield(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected yield for}

\item{aPeriod}{Current model period}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
AdaptiveExpectation_calcExpectedYield
}
\details{
Calculate the expected yield for a LandLeaf using
         an adaptive approach -- use linear combination of previous expectation and
         new information.
}
\author{
KVC November 2017
}
