% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{add_parameter_data}
\alias{add_parameter_data}
\title{Add model parameter values to a table of model results}
\usage{
add_parameter_data(df, aScenarioInfo)
}
\arguments{
\item{df}{Data frame with model results}

\item{aScenarioInfo}{ScenarioInfo structure for the scenario}
}
\description{
Add model parameter values to a table of model results
}
\keyword{internal}
