% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_production_technology.R
\name{AgProductionTechnology_calcProfitRate}
\alias{AgProductionTechnology_calcProfitRate}
\title{AgProductionTechnology_calcProfitRate}
\usage{
AgProductionTechnology_calcProfitRate(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{Land leaf}

\item{aPeriod}{Current model period}
}
\description{
AgProductionTechnology_calcProfitRate
}
\details{
Calculates the profit rate which is equal to the market price minus
         the variable cost. Profit rate can be negative.
         Profit rate is in 1975$ per billion m2, so computation includes yield.
}
\author{
KVC September 2017
}
