% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_production_technology.R
\name{AgProductionTechnology_readData}
\alias{AgProductionTechnology_readData}
\title{AgProductionTechnology_readData}
\usage{
AgProductionTechnology_readData(aLandLeaf)
}
\arguments{
\item{aLandLeaf}{Land leaf}
}
\description{
AgProductionTechnology_readData
}
\details{
Read in technology related calibration data
         (e.g., CalOutput, technical change, cost, etc.)
}
\author{
KVC September 2017
}
