% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{AgProductionTechnology_setup}
\alias{AgProductionTechnology_setup}
\title{AgProductionTechnology_setup}
\usage{
AgProductionTechnology_setup(aLandLeaf, aAgData, ascentype)
}
\arguments{
\item{aLandLeaf}{Land leaf}

\item{aAgData}{Agricultural technology data}

\item{ascentype}{Scenario type: either "Reference" or "Hindcast"}
}
\description{
AgProductionTechnology_setup
}
\details{
Setup technology (e.g., CalOutput, technical change, cost, etc.)
}
\author{
KVC October 2017
}
