% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aland_allocator_item.R
\name{ALandAllocatorItem_calculateShareWeights}
\alias{ALandAllocatorItem_calculateShareWeights}
\title{ALandAllocatorItem_calculateShareWeights}
\usage{
ALandAllocatorItem_calculateShareWeights(aRegionName, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aRegionName}{Region name}

\item{aChoiceFnAbove}{Type of logit}

\item{aPeriod}{Model time period}
}
\description{
ALandAllocatorItem_calculateShareWeights
}
\details{
Calculates share weights for land leafs and land nodes
}
\author{
KVC September 2017
}
