% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\name{as.ScenarioInfo}
\alias{as.ScenarioInfo}
\alias{as.ScenarioInfo.environment}
\alias{as.ScenarioInfo.list}
\title{Convert an object to a \code{ScenarioInfo} object}
\usage{
as.ScenarioInfo(object)

\method{as.ScenarioInfo}{environment}(object)

\method{as.ScenarioInfo}{list}(object)
}
\arguments{
\item{object}{Object to be tested or converted.}
}
\description{
Convert an object to a \code{ScenarioInfo} object
}
\section{Methods (by class)}{
\itemize{
\item \code{environment}: Convert an environment to a \code{ScenarioInfo}

\item \code{list}: Convert a list to a \code{ScenarioInfo}
}}

