% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_expectation.R
\name{calc_lagged_expectation}
\alias{calc_lagged_expectation}
\title{Compute autoregressive model for price expectation}
\usage{
calc_lagged_expectation(t, alpha, pricetable)
}
\arguments{
\item{t}{The current year, for which we wish to calculate the expected
price.}

\item{alpha}{Coefficient of previous year term in the autoregressive
series.}

\item{pricetable}{Table of price vs. year.  This series is assumed
not to have any gaps in it and to be in year order, but neither of
these conditions are checked.}
}
\description{
This function computes price expectation as an autoregressive
function of the price time series.
\deqn{
y_i = \alpha y_{i-1} + (1-\alpha) x_i
}
}
\details{
There is some dispute over what the final term in this series
should be.  In this implementation the final term is just like
all the others; therefore, for \eqn{\alpha = 0} this formula
reduces to \eqn{y_i = x_i}, the equivalent of perfect expectation.

This function is not currently vectorized; a separate call is needed
for each value of \code{t}.
}
