% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{ChoiceFunction}
\alias{ChoiceFunction}
\title{ChoiceFunction}
\usage{
ChoiceFunction(aType, aLogitExponent)
}
\arguments{
\item{aType}{Type of logit (currently only "relative-cost" is supported)}

\item{aLogitExponent}{Logt exponent}
}
\description{
ChoiceFunction
}
\details{
Logit choice function, both type and exponent
}
\author{
KVC September 2017
}
