% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{ev_single}
\alias{ev_single}
\title{Helper function for EV}
\usage{
ev_single(samples, modelgroup, reportvars, weighted, lp)
}
\arguments{
\item{samples}{Data frame of Monte Carlo samples for a single model.}

\item{modelgroup}{Vector of names of columns that define the model
groupings.  These will be included in the results along with the
\code{reportvars}.}

\item{reportvars}{Vector of names of variables for which to report
expectations.}

\item{weighted}{If \code{TRUE}, weight samples by their posterior
probabilities.}

\item{lp}{Name of column containing the log posterior.}
}
\description{
Helper function for EV
}
\keyword{internal}
