% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_price_data.R
\docType{data}
\name{EXPECTED_PRICES}
\alias{EXPECTED_PRICES}
\title{EXPECTED_PRICES}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 396 rows and 3 columns.}
\usage{
EXPECTED_PRICES
}
\description{
A tibble where expected prices can be stored
}
\author{
Kate Calvin
}
\keyword{datasets}
