% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{export_results}
\alias{export_results}
\title{Export scenario results as a csv file}
\usage{
export_results(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations.}
}
\value{
Table of model results.
}
\description{
Filter the scenario information object for a particular scenario and export its results
}
\author{
KVC
}
