% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{extract_pointwise_loglike}
\alias{extract_pointwise_loglike}
\title{Extract the pointwise log likelihood values from a \code{ScenarioInfo} structure}
\usage{
extract_pointwise_loglike(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{ScenarioInfo structure to extract from}
}
\value{
Matrix of log likelihood values, points in rows, parameter samples in
columns
}
\description{
One "scenario" will generally produce several samples because the likelihood
calculations for all of the \code{xi} values will be stored in a single
scenario object.
}
\keyword{internal}
