% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{extract_post_vector}
\alias{extract_post_vector}
\title{Extract the posterior vector for the samples in a \code{ScenarioInfo} object}
\usage{
extract_post_vector(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{ScenarioInfo object}
}
\value{
vector of log posterior values.
}
\description{
Extract the posterior vector for the samples in a \code{ScenarioInfo} object
}
\keyword{internal}
