% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FAO_land_history}
\alias{FAO_land_history}
\title{Historical land use data}
\format{A data frame with 14953 rows and 4 variables:
\describe{
\item{region}{GCAM region name.}
\item{GCAM_commodity}{GCAM commodity name.}
\item{year}{Year of the observation.}
\item{area}{Land area allocated to that commodity for the region and year.}
}}
\source{
Kate!
}
\usage{
FAO_land_history
}
\description{
This dataset contains land use by GCAM region and GCAM commodity for the
years 1961--2010
}
\keyword{datasets}
