% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_AgProdChange}
\alias{get_AgProdChange}
\title{get_AgProdChange}
\usage{
get_AgProdChange(ascentype, SUBREGION = NULL)
}
\arguments{
\item{ascentype}{Scenario type: either "Reference" or "Hindcast"}

\item{SUBREGION}{Subregion to read data for}
}
\value{
Tibble containing AgProdChange by commodity and year
}
\description{
get_AgProdChange
}
\details{
Read in AgProdChange for all periods and return them
}
\author{
KVC October 2017
}
