% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\name{get_hindcast_AgProdChange}
\alias{get_hindcast_AgProdChange}
\title{get_hindcast_AgProdChange}
\usage{
get_hindcast_AgProdChange(aScenType)
}
\arguments{
\item{aScenType}{Scenario type}
}
\value{
AgProdChange in historical period
}
\description{
get_hindcast_AgProdChange
}
\details{
Read in FAO production & ha by country and crop,
         Aggregate to GCAM regions and commodities,
         Compute & return AgProdChange in historical period
}
\author{
KVC October 2017
}
