% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\name{get_hindcast_prices}
\alias{get_hindcast_prices}
\title{get_hindcast_prices}
\usage{
get_hindcast_prices()
}
\value{
prices in historical period
}
\description{
get_hindcast_prices
}
\details{
Read in FAO prices by country and crop,
         Aggregate to GCAM regions and commodities,
}
\author{
KVC October 2017
}
