% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\name{get_historic_yield_ratios}
\alias{get_historic_yield_ratios}
\title{get_historic_yield_ratios}
\usage{
get_historic_yield_ratios()
}
\value{
YieldRatio in historical period
}
\description{
get_historic_yield_ratios
}
\details{
Read in FAO production & ha by country and crop,
         Aggregate to GCAM regions and commodities,
         Calculate ratio of yield to first GCAM historical year
}
\author{
KVC June 2019
}
