% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\name{get_historic_yields}
\alias{get_historic_yields}
\title{get_historic_yields}
\usage{
get_historic_yields()
}
\value{
YieldRatio in historical period
}
\description{
get_historic_yields
}
\details{
Read in FAO production & ha by country and crop,
         Aggregate to GCAM regions and commodities,
}
\author{
KVC June 2019
}
