% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{get_PCHES_results}
\alias{get_PCHES_results}
\title{get_PCHES_results}
\usage{
get_PCHES_results(aScenarioInfo, aYear)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations.}

\item{aYear}{Year to get results}
}
\value{
Table of model results.
}
\description{
Process results for PCHES coupling. Specifically, report the fractional change in land
from 2010 by crop and region
}
\author{
KVC
}
