% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_per_to_yr}
\alias{get_per_to_yr}
\title{get_per_to_yr}
\usage{
get_per_to_yr(aPer)
}
\arguments{
\item{aPer}{Period to convert}
}
\value{
Year
}
\description{
get_per_to_yr
}
\details{
Convert a period to a year
}
\author{
KVC October 2017
}
