% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{get_prices}
\alias{get_prices}
\title{get_prices}
\usage{
get_prices(aPeriod)
}
\arguments{
\item{aPeriod}{Current model period (used only to determine whether prices need to be processed or read)}
}
\value{
Tibble containing prices by commodity and year
}
\description{
get_prices
}
\details{
Read in prices for all periods and return them
}
\author{
KVC October 2017
}
