% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\name{get_saved_share_weights}
\alias{get_saved_share_weights}
\title{get_saved_share_weights}
\usage{
get_saved_share_weights()
}
\value{
Share weights as a named vector
}
\description{
Read in share weights from a file
}
\author{
KVC December 2018
}
