% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{get_scenario_land_data}
\alias{get_scenario_land_data}
\title{Load land use results for a list of already-run scenarios}
\usage{
get_scenario_land_data(aScenarioList)
}
\arguments{
\item{aScenarioList}{List of ScenarioInfo structures for the runs.}
}
\value{
Table with region, commodity, year, and area.
}
\description{
Fetch land use results and aggregate to region, commodity, year, and scenario.
}
\section{TODO}{


Currently the land use results don't have a region column, so we can't
aggregate by region (we have to assume that the results are all for a single
region).  Once we add a region column to the model output, we will want to
add region to the grouping variables for this function.
}

