% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_yr_to_per}
\alias{get_yr_to_per}
\title{get_yr_to_per}
\usage{
get_yr_to_per(aYear, scentype)
}
\arguments{
\item{aYear}{Year to convert}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Period number
}
\description{
get_yr_to_per
}
\details{
Convert a year to a period number
}
\author{
KVC October 2017
}
