% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{getNest}
\alias{getNest}
\title{Construct a table of land node nestings}
\usage{
getNest(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land allocator structure}
}
\description{
The table describes the graph of node relationships.  There is one row for
each edge in the graph.  The columns are "parent" and "node", and they
name the parent and child in the edge.
}
