% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getScenName}
\alias{getScenName}
\title{getScenName}
\usage{
getScenName(aScenName, aExpectation, aYears, aAgFor, aAgForNonPast, aCrop)
}
\arguments{
\item{aScenName}{Scenario base name}

\item{aExpectation}{Expectation type}

\item{aYears}{Years for Adaptive or Linear expectations (NULL for Perfect)}

\item{aAgFor}{Logit exponent for AgroForestLand}

\item{aAgForNonPast}{Logit exponent for AgroForestLand_NonPasture}

\item{aCrop}{Logit exponent for Cropland}
}
\value{
Scenario name
}
\description{
getScenName
}
\details{
Get the scenario name to use in ensemble runs
}
\author{
KVC November 2017
}
