% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getStartYear}
\alias{getStartYear}
\title{getStartYear}
\usage{
getStartYear(aScenType)
}
\arguments{
\item{aScenType}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Year
}
\description{
getStartYear
}
\details{
Get model start year
}
\author{
KVC October 2017
}
