% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefun.R
\name{grand_table_objective}
\alias{grand_table_objective}
\title{Organize a list of ScenarioInfo objects that have been analyzed with
objective functions into a grand table of parameters.}
\usage{
grand_table_objective(aScenarioList)
}
\arguments{
\item{aScenarioList}{List of ScenarioInfo structures}
}
\value{
Data frame containing model parameters and objective function values
}
\description{
The table produced includes the model parameters and all user-specified objective
function evaluations, for all of the models in the input list.
}
\details{
Before this function can run, the scenario list must have been run through
\code{\link{run_objective}} to calculate all objective function values.
}
\author{
ACS May 2020
}
