% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_expectation.R
\name{LaggedExpectation_calcExpectedYield}
\alias{LaggedExpectation_calcExpectedYield}
\title{LaggedExpectation_calcExpectedYield}
\usage{
LaggedExpectation_calcExpectedYield(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected yield for}

\item{aPeriod}{Current model period}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
LaggedExpectation_calcExpectedYield
}
\details{
Calculate the expected yield for a LandLeaf using
         a lagged approach -- use previous expectation + fraction
         of new information.
}
\author{
KVC November 2017
}
