% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator}
\alias{LandAllocator}
\title{LandAllocator}
\usage{
LandAllocator(aRegionName, aLogitExponent, aLandAllocation)
}
\arguments{
\item{aRegionName}{Region name}

\item{aLogitExponent}{Logit exponent of the top level of the land allocator}

\item{aLandAllocation}{Land allocation for this region}
}
\value{
New, initialized LandAllocator
}
\description{
LandAllocator
}
\details{
Initialize an Class called LandAllocator
}
\author{
KVC September 2017
}
