% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LandAllocator_addChild}
\alias{LandAllocator_addChild}
\title{LandAllocator_addChild}
\usage{
LandAllocator_addChild(aLandAllocator, aChild, aParentNames)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aChild}{Child (Land node or leaf) to insert}

\item{aParentNames}{Names of the parent nodes}
}
\description{
LandAllocator_addChild
}
\details{
Insert a land leaf into the land allocator
         Loop through all of the land allocator's children
         until we find the parent. Then, add the leaf.
}
\author{
KVC October 2017
}
