% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_addToNest}
\alias{LandAllocator_addToNest}
\title{LandAllocator_addToNest}
\usage{
LandAllocator_addToNest(aLandAllocator, aNest)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aNest}{Current nest}
}
\value{
Updated nest
}
\description{
LandAllocator_addToNest
}
\details{
Determine all of the parent/child relationships
         in the land allocator.
}
\author{
KVC October 2017
}
