% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calcFinalLandAllocation}
\alias{LandAllocator_calcFinalLandAllocation}
\title{LandAllocator_calcFinalLandAllocation}
\usage{
LandAllocator_calcFinalLandAllocation(aRegionName, aPeriod)
}
\arguments{
\item{aRegionName}{Region name.}

\item{aPeriod}{Model time period.}
}
\description{
LandAllocator_calcFinalLandAllocation
}
\details{
Initiates calculation of land allocation
         Calls calibrateLandAllocator, calcLandShares, and calcLandAllocation
}
\author{
KVC September 2017
}
