% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calcLandShares}
\alias{LandAllocator_calcLandShares}
\title{LandAllocator_calcLandShares}
\usage{
LandAllocator_calcLandShares(aRegionName, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aRegionName}{Region name.}

\item{aChoiceFnAbove}{Type of logit choice function for parent node.}

\item{aPeriod}{Model time period.}
}
\description{
LandAllocator_calcLandShares
}
\details{
Calculates land shares.
         Sets unmanaged profit rate, then calls land node's
         calcLandShares. Also, sets root share to 1.
}
\author{
KVC September 2017
}
