% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calculateNodeProfitRates}
\alias{LandAllocator_calculateNodeProfitRates}
\title{LandAllocator_calculateNodeProfitRates}
\usage{
LandAllocator_calculateNodeProfitRates(aLandAllocator, aUnmanagedLandValue,
  aChoiceFunction, aPeriod)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aUnmanagedLandValue}{Value of unmanaged land}

\item{aChoiceFunction}{Choice function}

\item{aPeriod}{Current model period}
}
\description{
LandAllocator_calculateNodeProfitRates
}
\details{
Loops through all of the children of the aLandAllocator
         and calls the calculateNodeProfitRates method on each
}
\author{
KVC October 2017
}
