% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calculateShareWeights}
\alias{LandAllocator_calculateShareWeights}
\title{LandAllocator_calculateShareWeights}
\usage{
LandAllocator_calculateShareWeights(
  aLandAllocator,
  aChoiceFunction,
  aPeriod,
  aScenarioInfo
)
}
\arguments{
\item{aLandAllocator}{Land Allocator}

\item{aChoiceFunction}{Choice function}

\item{aPeriod}{Current time period}

\item{aScenarioInfo}{Scenario info object}
}
\description{
LandAllocator_calculateShareWeights
}
\details{
Loop through all children of the land allocator and
         call the calculateShareWeights method on each
}
\author{
KVC October 2017
}
