% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calibrateLandAllocator}
\alias{LandAllocator_calibrateLandAllocator}
\title{LandAllocator_calibrateLandAllocator}
\usage{
LandAllocator_calibrateLandAllocator(aLandAllocator, aPeriod)
}
\arguments{
\item{aLandAllocator}{Land allocator for this region}

\item{aPeriod}{model period.}
}
\description{
LandAllocator_calibrateLandAllocator
}
\details{
Sets initial land shares, sets unmanaged land profit rates,
         calculate calibration prfot rates, then calculates share profit scalers
         that will be used for scaling the profit rates for future year sharing
}
\author{
KVC September 2017
}
