% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getExpectedPrice}
\alias{LandAllocator_getExpectedPrice}
\title{LandAllocator_getExpectedPrice}
\usage{
LandAllocator_getExpectedPrice(aLandAllocator, aData)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aData}{Data table to store expected price}
}
\value{
Expected price data
}
\description{
LandAllocator_getExpectedPrice
}
\details{
Calculates and returns expected price for all leafs
}
\author{
KVC October 2017
}
