% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getExpectedYield}
\alias{LandAllocator_getExpectedYield}
\title{LandAllocator_getExpectedYield}
\usage{
LandAllocator_getExpectedYield(aLandAllocator, aName, aPeriod)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aName}{Name of leaf we want expected yield for}

\item{aPeriod}{Model period}
}
\value{
Expected yield
}
\description{
LandAllocator_getExpectedYield
}
\details{
Calculates and returns expected yield for a particular leaf
}
\author{
KVC October 2017
}
