% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LandAllocator_getLandAreaByCriteria}
\alias{LandAllocator_getLandAreaByCriteria}
\title{LandAllocator_getLandAreaByCriteria}
\usage{
LandAllocator_getLandAreaByCriteria(aLandAllocator, aString, aPeriod)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aString}{String to group land area by}

\item{aPeriod}{Model period to get land area for}
}
\value{
Land area
}
\description{
LandAllocator_getLandAreaByCriteria
}
\details{
Calculate total land area that meets a particular criteria.
         Criteria is currently determined by a string appearing in the
         name of the land type
}
