% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getLandShares}
\alias{LandAllocator_getLandShares}
\title{LandAllocator_getLandShares}
\usage{
LandAllocator_getLandShares(aLandAllocator, aShares)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aShares}{Table of shares to append information to}
}
\value{
Table of land shares
}
\description{
LandAllocator_getLandShares
}
\details{
Calculates and returns land allocation for a particular leaf
}
\author{
KVC November 2017
}
