% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getOutputs}
\alias{LandAllocator_getOutputs}
\title{LandAllocator_getOutputs}
\usage{
LandAllocator_getOutputs(aLandAllocator, aAllOutputs, aScenType)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aAllOutputs}{Data frame to fill in outputs}

\item{aScenType}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Output table
}
\description{
LandAllocator_getOutputs
}
\details{
Calculates and returns all outputs for the land allocator
}
\author{
KVC May 2018
}
