% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getYield}
\alias{LandAllocator_getYield}
\title{LandAllocator_getYield}
\usage{
LandAllocator_getYield(aLandAllocator, aData)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aData}{Table to store data in}
}
\value{
Yield for all leafs
}
\description{
LandAllocator_getYield
}
\details{
Calculates and returns yield for all leafs
}
\author{
KVC October 2017
}
