% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_readData}
\alias{LandAllocator_readData}
\title{LandAllocator_readData}
\usage{
LandAllocator_readData(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs data}
}
\description{
LandAllocator_readData
}
\details{
Read in calibration data for the land allocator
}
\author{
KVC September 2017
}
