% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_setInitShares}
\alias{LandAllocator_setInitShares}
\title{LandAllocator_setInitShares}
\usage{
LandAllocator_setInitShares(aLandAllocator, aPeriod)
}
\arguments{
\item{aLandAllocator}{Land allocator for this region.}

\item{aPeriod}{model period.}
}
\description{
LandAllocator_setInitShares
}
\details{
Calculates and sets initial land shares
}
\author{
KVC September 2017
}
