% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_setUnmanagedLandProfitRate}
\alias{LandAllocator_setUnmanagedLandProfitRate}
\title{LandAllocator_setUnmanagedLandProfitRate}
\usage{
LandAllocator_setUnmanagedLandProfitRate(
  aLandAllocator,
  aUnmanagedLandValue,
  aPeriod
)
}
\arguments{
\item{aLandAllocator}{Current land allocator}

\item{aUnmanagedLandValue}{Value of unmanaged land}

\item{aPeriod}{Current period}
}
\description{
LandAllocator_setUnmanagedLandProfitRate
}
\details{
Initalize setting the unmanaged land value in all unmanaged land leafs,
         by calling `setUnmanagedLandProfitRate` on all children
}
\author{
KVC October 2017
}
