% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LandAllocator_setup}
\alias{LandAllocator_setup}
\title{LandAllocator_setup}
\usage{
LandAllocator_setup(aLandAllocator, aScenarioInfo, agData = NULL,
  SUBREGION = NULL)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{agData}{Ag data read by \code{\link{ReadData_AgProd}}.  The data must
be for the same region and scenario type as the \code{aScenarioInfo} object.}

\item{SUBREGION}{Subregion to read data for}
}
\description{
LandAllocator_setup
}
\details{
Setup the land allocator. This includes
         Reading in external data and organizing
         it in the node, leaf, technology structure
}
\author{
KVC October 2017
}
