% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_calcLandAllocation}
\alias{LandLeaf_calcLandAllocation}
\title{LandLeaf_calcLandAllocation}
\usage{
LandLeaf_calcLandAllocation(aLandLeaf, aLandAllocationAbove, aPeriod)
}
\arguments{
\item{aLandLeaf}{Land leaf to perform calculations on}

\item{aLandAllocationAbove}{Land allocated to the parent node}

\item{aPeriod}{Model period}
}
\description{
LandLeaf_calcLandAllocation
}
\details{
Land allocation is the product of the land
         allocated to the parent node and the share
         of land specified for this land leaf which is calculated
         previously using the logit function. Note: this method
         is called in every time period including calibration
         periods. Thus, land in a calibration period is not
         necessarily equal to read in values
}
\author{
KVC September 2017
}
