% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_calcLandShares}
\alias{LandLeaf_calcLandShares}
\title{LandLeaf_calcLandShares}
\usage{
LandLeaf_calcLandShares(aLandLeaf, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{The discrete choice function from the level above.}

\item{aPeriod}{Model period}

\item{aRegionName}{Region.}
}
\value{
Unormalized land shares
}
\description{
LandLeaf_calcLandShares
}
\details{
This method implements the logit function. A land type's
         share of land is based on its profit rate and the
         distribution assumed for the parent node ( aLogitExpAbove )
}
\author{
KVC September 2017
}
