% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_calculateShareWeights}
\alias{LandLeaf_calculateShareWeights}
\title{LandLeaf_calculateShareWeights}
\usage{
LandLeaf_calculateShareWeights(aLandLeaf, aChoiceFnAbove, aPeriod, aParent,
  aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{Land leaf to perform calculations on}

\item{aChoiceFnAbove}{Type of logit}

\item{aPeriod}{Model time period}

\item{aParent}{Parent node (needed for calibrating new techs)}

\item{aScenarioInfo}{Scenario info object}
}
\description{
LandLeaf_calculateShareWeights
}
\details{
Calculates share weights for land leafs
}
\author{
KVC September 2017
}
