% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_getCalLandAllocation}
\alias{LandLeaf_getCalLandAllocation}
\title{LandLeaf_getCalLandAllocation}
\usage{
LandLeaf_getCalLandAllocation(aLandLeaf, aPeriod)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{aPeriod}{Model Period}
}
\value{
Land allocation for this leaf in this time period
}
\description{
LandLeaf_getCalLandAllocation
}
\details{
Get the amount of land in a particular period
}
\author{
KVC October 2017
}
