% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandLeaf_getExpectedPrice}
\alias{LandLeaf_getExpectedPrice}
\title{LandLeaf_getExpectedPrice}
\usage{
LandLeaf_getExpectedPrice(aLandLeaf, aData, scentype)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{aData}{Table to store data in}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Expected price data
}
\description{
LandLeaf_getExpectedPrice
}
\details{
Calculates and returns expected price for this leaf
}
\author{
KVC November 2017
}
