% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandLeaf_getExpectedYield}
\alias{LandLeaf_getExpectedYield}
\title{LandLeaf_getExpectedYield}
\usage{
LandLeaf_getExpectedYield(aLandLeaf, aData, scentype)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{aData}{Table to store data in}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Expected yield data
}
\description{
LandLeaf_getExpectedYield
}
\details{
Calculates and returns expected yield for this leaf
}
\author{
KVC October 2017
}
