% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LandLeaf_getLandAreaByCriteria}
\alias{LandLeaf_getLandAreaByCriteria}
\title{LandLeaf_getLandAreaByCriteria}
\usage{
LandLeaf_getLandAreaByCriteria(aLandLeaf, aString, aPeriod)
}
\arguments{
\item{aLandLeaf}{Land allocator}

\item{aString}{String to group land area by}

\item{aPeriod}{Model period to get land area for}
}
\value{
Land area for this leaf
}
\description{
LandLeaf_getLandAreaByCriteria
}
\details{
Calculate total land area that meets a particular criteria.
         Criteria is currently determined by a string appearing in the
         name of the land type
}
