% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandLeaf_getOutputs}
\alias{LandLeaf_getOutputs}
\title{LandLeaf_getOutputs}
\usage{
LandLeaf_getOutputs(aLandLeaf, aAllOutputs, aScenType)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{aAllOutputs}{Data frame to fill in outputs}

\item{aScenType}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Output table
}
\description{
LandLeaf_getOutputs
}
\details{
Calculates and returns outputs for this leaf in all periods
}
\author{
KVC May 2018
}
