% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandLeaf_getYield}
\alias{LandLeaf_getYield}
\title{LandLeaf_getYield}
\usage{
LandLeaf_getYield(aLandLeaf, aData)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{aData}{Table to store data in}
}
\value{
Yield for all leafs in the node
}
\description{
LandLeaf_getYield
}
\details{
Calculates and returns yield for this leaf
}
\author{
KVC October 2017
}
