% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_setInitShares}
\alias{LandLeaf_setInitShares}
\title{LandLeaf_setInitShares}
\usage{
LandLeaf_setInitShares(aLandLeaf, aLandAllocationAbove, aPeriod)
}
\arguments{
\item{aLandAllocationAbove}{Land allocation of the parent node}

\item{aPeriod}{Model period}

\item{aRegionName}{Region.}
}
\description{
LandLeaf_setInitShares
}
\details{
Calculates the share of land allocated to a leaf.
         This method is called during the calibration process
         so the shares set are prior to any calculations
         of share weights.
}
\author{
KVC September 2017
}
