% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode}
\alias{LandNode}
\title{LandNode}
\usage{
LandNode(aName, aLogitExponent, aLandAllocation)
}
\arguments{
\item{aName}{Node name}

\item{aLogitExponent}{Logit exponent of the node}

\item{aLandAllocation}{Land allocation for this node}
}
\value{
New, initialized LandNode
}
\description{
LandNode
}
\details{
Initialize an Class called LandNode
}
\author{
KVC September 2017
}
