% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LandNode_addChild}
\alias{LandNode_addChild}
\title{LandNode_addChild}
\usage{
LandNode_addChild(aLandNode, aChild, aParentNames)
}
\arguments{
\item{aLandNode}{Current land node}

\item{aChild}{Child (land node or leaf) to insert}

\item{aParentNames}{Names of the parent nodes}
}
\description{
LandNode_addChild
}
\details{
Insert a land leaf or node into the land allocator
         Loop through all of this land node's children
         until we find the parent. Then, add the leaf or node.
}
\author{
KVC October 2017
}
