% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_addToNest}
\alias{LandNode_addToNest}
\title{LandNode_addToNest}
\usage{
LandNode_addToNest(aLandNode, aNest)
}
\arguments{
\item{aLandNode}{Land node}

\item{aNest}{Current nest}
}
\value{
Updated nest
}
\description{
LandNode_addToNest
}
\details{
Determine all of the parent/child relationships
         for this particular node.
}
\author{
KVC October 2017
}
