% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_calcLandAllocation}
\alias{LandNode_calcLandAllocation}
\title{LandNode_calcLandAllocation}
\usage{
LandNode_calcLandAllocation(aRegionName, aLandAllocationAbove, aPeriod)
}
\arguments{
\item{aRegionName}{Region name.}

\item{aLandAllocationAbove}{Land allocation of parent.}

\item{aPeriod}{model period.}
}
\description{
LandNode_calcLandAllocation
}
\details{
Uses the land share and the allocation of land to
         the parent node to calculate the allocation of this
         node. CalculateLandShares must be called first.
}
\author{
KVC September 2017
}
