% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_calculateNodeProfitRates}
\alias{LandNode_calculateNodeProfitRates}
\title{LandNode_calculateNodeProfitRates}
\usage{
LandNode_calculateNodeProfitRates(aLandNode, aAverageProfitRateAbove,
  aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aAverageProfitRateAbove}{Average profit rate of the parent node.}

\item{aChoiceFnAbove}{The discrete choice function from the level above.}

\item{aPeriod}{Period.}

\item{aRegionName}{Region}
}
\description{
LandNode_calculateNodeProfitRates
}
\details{
Calculates profit rates at the node level using the
         logit function calculations.
}
\author{
KVC September 2017
}
