% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_calculateShareWeights}
\alias{LandNode_calculateShareWeights}
\title{LandNode_calculateShareWeights}
\usage{
LandNode_calculateShareWeights(aRegionName, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aRegionName}{Region}

\item{aChoiceFnAbove}{The discrete choice function from the level above.}

\item{aPeriod}{Period.}
}
\description{
LandNode_calculateShareWeights
}
\details{
Calculates share weights at the node level and calls
         the share weight calculation for the node's children.
}
\author{
KVC September 2017
}
