% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_getCalLandAllocation}
\alias{LandNode_getCalLandAllocation}
\title{LandNode_getCalLandAllocation}
\usage{
LandNode_getCalLandAllocation(aLandNode, aPeriod)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aPeriod}{Model period}
}
\description{
LandNode_getCalLandAllocation
}
\details{
Calculates and returns total land allocation of a given type.
}
\author{
KVC October 2017
}
