% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_getChildWithHighestShare}
\alias{LandNode_getChildWithHighestShare}
\title{LandNode_getChildWithHighestShare}
\usage{
LandNode_getChildWithHighestShare(aPeriod)
}
\arguments{
\item{aPeriod}{Model time period.}
}
\description{
LandNode_getChildWithHighestShare
}
\details{
Finds the child with the highest share (used for bioenergy calibration)
}
\author{
KVC September 2017
}
