% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getExpectedPrice}
\alias{LandNode_getExpectedPrice}
\title{LandNode_getExpectedPrice}
\usage{
LandNode_getExpectedPrice(aLandNode, aData)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aData}{Data table to store expected price}
}
\value{
Expected price data
}
\description{
LandNode_getExpectedPrice
}
\details{
Calculates and returns expected price for leafs in this node
}
\author{
KVC November 2017
}
