% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getExpectedYield}
\alias{LandNode_getExpectedYield}
\title{LandNode_getExpectedYield}
\usage{
LandNode_getExpectedYield(aLandNode, aName, aPeriod)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aName}{Name of leaf we want expected yield for}

\item{aPeriod}{Model period}
}
\value{
Expected yield for leaf named `aName`
}
\description{
LandNode_getExpectedYield
}
\details{
Calculates and returns expected yield for a given leaf
}
\author{
KVC October 2017
}
