% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getLandAllocation}
\alias{LandNode_getLandAllocation}
\title{LandNode_getLandAllocation}
\usage{
LandNode_getLandAllocation(aLandNode, allLand, scentype)
}
\arguments{
\item{aLandNode}{LandNode}

\item{allLand}{Data frame to fill in land allocation}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Land allocation table
}
\description{
LandNode_getLandAllocation
}
\details{
Calculates and returns total land allocation for types and periods
}
\author{
KVC October 2017
}
