% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LandNode_getLandAreaByCriteria}
\alias{LandNode_getLandAreaByCriteria}
\title{LandNode_getLandAreaByCriteria}
\usage{
LandNode_getLandAreaByCriteria(aLandNode, aString, aPeriod)
}
\arguments{
\item{aLandNode}{Land allocator}

\item{aString}{String to group land area by}

\item{aPeriod}{Model period to get land area for}
}
\value{
Land area
}
\description{
LandNode_getLandAreaByCriteria
}
\details{
Calculate total land area that meets a particular criteria.
         Criteria is currently determined by a string appearing in the
         name of the land type
}
