% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_getObservedAverageProfitRate}
\alias{LandNode_getObservedAverageProfitRate}
\title{LandNode_getObservedAverageProfitRate}
\usage{
LandNode_getObservedAverageProfitRate(aProfitRate, aShare, aPeriod)
}
\arguments{
\item{aProfitRate}{Profit rate.}

\item{aShare}{Share.}

\item{aPeriod}{Model time period.}
}
\description{
LandNode_getObservedAverageProfitRate
}
\details{
Get observed average profit rate.
}
\author{
KVC September 2017
}
