% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getOutputs}
\alias{LandNode_getOutputs}
\title{LandNode_getOutputs}
\usage{
LandNode_getOutputs(aLandNode, aAllOutputs, aScenType)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aAllOutputs}{Data frame to fill in outputs}

\item{aScenType}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Output table
}
\description{
LandNode_getOutputs
}
\details{
Calculates and returns all outputs for this node
}
\author{
KVC May 2018
}
