% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getYield}
\alias{LandNode_getYield}
\title{LandNode_getYield}
\usage{
LandNode_getYield(aLandNode, aData, scentype)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aData}{Table to store data in}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Yield for all leafs in the node
}
\description{
LandNode_getYield
}
\details{
Calculates and returns yield for all leafs in a node
}
\author{
KVC October 2017
}
