% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_initCalc}
\alias{LandNode_initCalc}
\title{LandNode_initCalc}
\usage{
LandNode_initCalc(aLandNode, aPeriod)
}
\arguments{
\item{aLandNode}{Land node.}

\item{aPeriod}{Model time period.}
}
\description{
LandNode_initCalc
}
\details{
Initial calculations for the land node.
         Currently, this calls initCalc on children.
}
\author{
KVC September 2017
}
