% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_setInitShares}
\alias{LandNode_setInitShares}
\title{LandNode_setInitShares}
\usage{
LandNode_setInitShares(aLandNode, aLandAllocationAbove, aPeriod)
}
\arguments{
\item{aLandNode}{Land node to perform calculations on}

\item{aLandAllocationAbove}{Land allocation of the parent node}

\item{aPeriod}{Model period}
}
\description{
LandNode_setInitShares
}
\details{
Calculates the share of land allocated to a node and calls
         a similar method for the node's children. This method is
         called during the calibration process so the shares
         set are prior to any calculations of share weights.
}
\author{
KVC September 2017
}
