% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{Leaf_setup}
\alias{Leaf_setup}
\title{Leaf_setup}
\usage{
Leaf_setup(aLandAllocator, aRegionName, data, col.name)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}
}
\description{
Leaf_setup
}
\details{
Setup land leafs for the land allocator. This includes
         searching for the right parent, creating and creating leafs.
}
\author{
KVC October 2017
}
