% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_expectation.R
\name{LinearExpectation_calcExpectedPrice}
\alias{LinearExpectation_calcExpectedPrice}
\title{LinearExpectation_calcExpectedPrice}
\usage{
LinearExpectation_calcExpectedPrice(aLandLeaf, aPeriod)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected price for}

\item{aPeriod}{Current model period}
}
\description{
LinearExpectation_calcExpectedPrice
}
\details{
Calculate the expected price for a LandLeaf using
         a linear extrapolation from recent history.
}
\author{
KVC October 2017
}
