% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_expectation.R
\name{LinearExpectation_calcExpectedYield}
\alias{LinearExpectation_calcExpectedYield}
\title{LinearExpectation_calcExpectedYield}
\usage{
LinearExpectation_calcExpectedYield(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected yield for}

\item{aPeriod}{Current model period}
}
\description{
LinearExpectation_calcExpectedYield
}
\details{
Calculate the expected yield for a LandLeaf using
         a linear extrapolation from recent history.
}
\author{
KVC October 2017
}
