% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LN1_setup}
\alias{LN1_setup}
\title{LN1_setup}
\usage{
LN1_setup(aLandAllocator, aRegionName, aData, aColName, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}

\item{aRegionName}{Region}

\item{aData}{Data needed for setting up this node}

\item{aColName}{Column name with the parent}
}
\description{
LN1_setup
}
\details{
Setup LN1 of the land allocator. This includes
         reading in external data, creating LandNode1 nodes, and
         creating any leafs with only one parent node.
}
\author{
KVC October 2017
}
