% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{log_sum}
\alias{log_sum}
\title{Compute the log of the sum of values given as logs}
\usage{
log_sum(lx)
}
\arguments{
\item{lx}{Vector of logs}
}
\value{
log of the sum of the antilogs of the input values.
}
\description{
Compute \code{log(sum(exp(x)))}, but do so in a way that avoids underflow.
}
\details{
Use the following identity to stabilize the calculation:
\deqn{\sum \exp(x-x_0) = 1/x0 \sum \exp(x)}
}
