% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{lppd}
\alias{lppd}
\title{Compute the log pointwise predictive density for a single model.}
\usage{
lppd(pointwise_ll)
}
\arguments{
\item{pointwise_ll}{List returned from \code{\link{pointwise_loglike}}.}
}
\value{
A vector of lppd values averaged over Monte Carlo samples.  Each
observation will have one corresponding pont in the vector.
}
\description{
Compute the log pointwise predictive density for a single model.
}
\keyword{internal}
