% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{lppdavg}
\alias{lppdavg}
\title{Compute the log average pointwise predictive density for a single model.}
\usage{
lppdavg(pointwise_ll)
}
\arguments{
\item{pointwise_ll}{List returned from \code{\link{pointwise_loglike}}.}
}
\description{
Compute the log average pointwise predictive density for a single model.
}
\keyword{internal}
