% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefun.R
\name{MAP_objective}
\alias{MAP_objective}
\title{Calculate parameters than minimize objective function of choice for a collection of model runs}
\usage{
MAP_objective(samples, modelgroup = "expectation.type",
  reportvars = NULL, objfun_to_min = "rms", landtypes = NULL)
}
\arguments{
\item{samples}{Monte Carlo samples, given either as a grand table or a list
of \code{ScenarioInfo} objects}

\item{modelgroup}{Vector of names of columns that define the model groupings.
The default is the single column \code{expectation.type}.}

\item{reportvars}{Vector of names of variables for which to report
expectations.  The default is all parameter values.}

\item{objfun_to_min}{a single objective function to minimize, from the supported
c('bias', 'rms', 'centeredrms', 'nrms', 'centerednrms', 'kge')). Defaults to rms.
Must be of length 1.}

\item{landtypes}{Vector of land types to include in minimization. (default is
to use  \code{c( "Corn", "FiberCrop", "MiscCrop", "OilCrop",
 "OtherGrain", "PalmFruit",  "Rice", "Root_Tuber","SugarCrop",  "Wheat"))}.}
}
\description{
Scans the runs in a given sample of model runs and selects the parameter set from those runs
that minimizes a specified objective function across specified land.types.
The final minimized value is currently coded as the mean across specified land.types of the
specified objective function of interest.
}
\details{
TODO: update the name so that it's not MAP anymore; currently naming this way for workflow
consistency with bayesian analysis, but we aren't doing Maximum A Posteriori analysis anymore,
so MAP is kind of misleading. I don't really have a better name right now though.

TODO: more detail on calculation being done.
}
\author{
ACS May 2020
}
