% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{outdir_setup}
\alias{outdir_setup}
\title{Set up output directories}
\usage{
outdir_setup(outdir)
}
\arguments{
\item{outdir}{Name of the output directory}
}
\value{
Normalized name of the output directory (invisibly)
}
\description{
Check the existence of the output directory (supplied as an argument) and its
required subdirectories (generated internally).  If any do not exist, create them.
Failure to create any of the required directories is a fatal error.
}
