% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_expectation.R
\name{PerfectExpectation_calcExpectedPrice}
\alias{PerfectExpectation_calcExpectedPrice}
\title{PerfectExpectation_calcExpectedPrice}
\usage{
PerfectExpectation_calcExpectedPrice(aLandLeaf, aPeriod, aScenarioInfo)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected price for}

\item{aPeriod}{Current model period}

\item{aScenarioInfo}{Scenario parameter structure}
}
\description{
PerfectExpectation_calcExpectedPrice
}
\details{
Calculate the expected price for a LandLeaf assuming
         perfect expectations, i.e., expected = actual
}
\author{
KVC October 2017
}
