% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_expectation.R
\name{PerfectExpectation_calcExpectedYield}
\alias{PerfectExpectation_calcExpectedYield}
\title{PerfectExpectation_calcExpectedYield}
\usage{
PerfectExpectation_calcExpectedYield(aLandLeaf, aPeriod)
}
\arguments{
\item{aLandLeaf}{LandLeaf to calculate expected yield for}

\item{aPeriod}{Current model period}
}
\description{
PerfectExpectation_calcExpectedYield
}
\details{
Calculate the expected yield assuming perfect
         information (i.e., expected = actual).
}
\author{
KVC October 2017
}
