% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plot_LandAllocation}
\alias{plot_LandAllocation}
\title{plot_LandAllocation}
\usage{
plot_LandAllocation(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land Allocator}
}
\description{
plot_LandAllocation
}
\details{
Plots land allocation over time
}
\author{
KVC September 2017
}
