% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plot_Nest}
\alias{plot_Nest}
\title{plot_Nest}
\usage{
plot_Nest(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land allocator}
}
\description{
plot_Nest
}
\details{
Plot the nesting structure for the land allocator
}
\author{
KVC October 2017
}
