% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plot_RegionalLandAllocation}
\alias{plot_RegionalLandAllocation}
\title{plot_RegionalLandAllocation}
\usage{
plot_RegionalLandAllocation(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land Allocator}
}
\description{
plot_RegionalLandAllocation
}
\details{
Plots regional land allocation over time.
         Aggregates land by region (i.e., adds all AEZs together)
         and then plots land.
}
\author{
KVC September 2017

KVC October 2017
}
