% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plotNest}
\alias{plotNest}
\title{plotNest}
\usage{
plotNest(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
plotNest
}
\details{
Plot the nesting structure for the land allocator
}
\author{
KVC October 2017
}
