% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_price_data.R
\docType{data}
\name{PRICES}
\alias{PRICES}
\title{PRICES}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 192 rows and 4 columns.}
\usage{
PRICES
}
\description{
A tibble with all of the prices for the model
}
\author{
Kate Calvin
}
\keyword{datasets}
