% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printDebug}
\alias{printDebug}
\title{Write out additional information for debugging}
\usage{
printDebug(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator structure}

\item{aScenarioInfo}{Scenario parameter structure}
}
\description{
Write the land shares and land nests to their respective output files.  This
information will be written out if the \code{aVerbose} argument to
\code{\link{run_model}} is \code{TRUE}.
}
