% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printExpectedPrice}
\alias{printExpectedPrice}
\title{printExpectedPrice}
\usage{
printExpectedPrice(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
printExpectedPrice
}
\details{
Prints expected price by land leaf
}
\author{
KVC October 2017
}
