% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printExpectedPrices}
\alias{printExpectedPrices}
\title{printExpectedPrices}
\usage{
printExpectedPrices(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
printExpectedPrices
}
\details{
Print expected prices
}
\author{
KVC November 2017
}
