% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printExpectedYield}
\alias{printExpectedYield}
\title{printExpectedYield}
\usage{
printExpectedYield(aLandAllocator, aScenarioInfo, aNest)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{aNest}{Land allocator nest}
}
\description{
printExpectedYield
}
\details{
Prints expected yield by land leaf
}
\author{
KVC October 2017
}
