% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printLandAllocation}
\alias{printLandAllocation}
\title{printLandAllocation}
\usage{
printLandAllocation(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator}
}
\description{
printLandAllocation
}
\details{
Prints land allocation by land leaf
}
\author{
KVC October 2017
}
