% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printLandShares}
\alias{printLandShares}
\title{printLandShares}
\usage{
printLandShares(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land allocator}
}
\description{
printLandShares
}
\details{
Prints land share for all nodes and leafs
}
\author{
KVC November 2017
}
