% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printNest}
\alias{printNest}
\title{printNest}
\usage{
printNest(aLandAllocator)
}
\arguments{
\item{aLandAllocator}{Land allocator}
}
\description{
printNest
}
\details{
Prints the land allocator nest, which can
         be used to plot tree structure or to query
         leaf data
}
\author{
KVC October 2017
}
