% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printPrices}
\alias{printPrices}
\title{printPrices}
\usage{
printPrices(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
printPrices
}
\details{
Print prices
}
\author{
KVC November 2017
}
