% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printYield}
\alias{printYield}
\title{printYield}
\usage{
printYield(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
printYield
}
\details{
Prints yield by land leaf
}
\author{
KVC October 2017
}
