% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{pwaic}
\alias{pwaic}
\title{Compute the effective number of parameters \code{pwaic} for a model.}
\usage{
pwaic(pointwise_ll)
}
\arguments{
\item{pointwise_ll}{List returned from \code{\link{pointwise_loglike}}.}
}
\description{
For each observation, compute the variance of the log likelihood.  Then,
\code{pwaic} is the sum of these.  Note that we are computing the variance of
the \emph{log} likelihood
}
\keyword{internal}
