% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_AgProd}
\alias{ReadData_AgProd}
\title{ReadData_AgProd}
\usage{
ReadData_AgProd(aRegionName)
}
\arguments{
\item{aRegionName}{Region to read data for}
}
\value{
All AgProductionTechnology information
}
\description{
ReadData_AgProd
}
\details{
Read in ag production data
}
\author{
KVC October 2017
}
