% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LandAllocator}
\alias{ReadData_LandAllocator}
\title{ReadData_LandAllocator}
\usage{
ReadData_LandAllocator()
}
\value{
Land allocator data
}
\description{
ReadData_LandAllocator
}
\details{
Read in total land allocation and logit exponents
         for the LandAllocator.
}
\author{
KVC October 2017
}
