% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN0}
\alias{ReadData_LN0}
\title{ReadData_LN0}
\usage{
ReadData_LN0(aRegionName, SUBREGION = NULL, subregionData = NULL)
}
\arguments{
\item{aRegionName}{Region to read data for}

\item{SUBREGION}{Subregion to read data for}

\item{subregionData}{Subregion data}
}
\value{
Land allocator data
}
\description{
ReadData_LN0
}
\details{
Read in total land allocation and logit exponents
         for the LandAllocator.
}
\author{
KVC October 2017
}
