% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN2_Node}
\alias{ReadData_LN2_Node}
\title{ReadData_LN2_Node}
\usage{
ReadData_LN2_Node(aRegionName)
}
\arguments{
\item{aRegionName}{Region to read data for}
}
\value{
Data on level 2 nodes of the land allocator
}
\description{
ReadData_LN2_Node
}
\details{
Read in unmanaged land value, names of children, and logit exponents
         for the level 2 nodes
}
\author{
KVC October 2017
}
