% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN3_GhostShare}
\alias{ReadData_LN3_GhostShare}
\title{ReadData_LN3_GhostShare}
\usage{
ReadData_LN3_GhostShare(aRegionName)
}
\arguments{
\item{aRegionName}{Region to read data for}
}
\value{
Ghost shares for level 3 nodes
}
\description{
ReadData_LN3_GhostShare
}
\details{
Read in ghost share for LN3 nodes
}
\author{
KVC October 2017
}
