% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN3_LandLeaf}
\alias{ReadData_LN3_LandLeaf}
\title{ReadData_LN3_LandLeaf}
\usage{
ReadData_LN3_LandLeaf(aRegionName, SUBREGION = NULL,
  subregionData = NULL, aScenarioInfo)
}
\arguments{
\item{aRegionName}{Region to read data for}

\item{SUBREGION}{Subregion to read data for}

\item{subregionData}{Subregion data}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\value{
Data on LandLeaf children of LandNode2
}
\description{
ReadData_LN3_LandLeaf
}
\details{
Read in names of information on LandLeafs that are
         children of LandNode3 nodes
}
\author{
KVC October 2017
}
