% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN3_NewTech}
\alias{ReadData_LN3_NewTech}
\title{ReadData_LN3_NewTech}
\usage{
ReadData_LN3_NewTech(aRegionName)
}
\arguments{
\item{aRegionName}{Region to read data for}
}
\value{
Data on new technologies for children of LandNode3
}
\description{
ReadData_LN3_NewTech
}
\details{
Read in new technologies that are children of LandNode3 nodes
}
\author{
KVC May 2018
}
