% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN3_UnmanagedLandLeaf}
\alias{ReadData_LN3_UnmanagedLandLeaf}
\title{ReadData_LN3_UnmanagedLandLeaf}
\usage{
ReadData_LN3_UnmanagedLandLeaf(aRegionName, aSubRegion, subregionData)
}
\arguments{
\item{aRegionName}{Region to read data for}

\item{aSubRegion}{Subregion to read data for}

\item{subregionData}{Subregion data}
}
\value{
Data on UnmanagedLandLeaf children of LandNode3
}
\description{
ReadData_LN3_UnmanagedLandLeaf
}
\details{
Read in names of information on UnmanagedLandLeafs that are
         children of LandNode3 nodes
}
\author{
KVC October 2017
}
