% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcAverageCost}
\alias{RelativeCostLogit_calcAverageCost}
\title{RelativeCostLogit_calcAverageCost}
\usage{
RelativeCostLogit_calcAverageCost(aChoiceFunction, aUnnormalizedShareSum,
  aLogShareFac, aPeriod)
}
\arguments{
\item{aChoiceFunction}{Choice function}

\item{aUnnormalizedShareSum}{Sum of unnormalized shares of children}

\item{aLogShareFac}{lfac}

\item{aPeriod}{Current model period}
}
\value{
Profit rate of the node
}
\description{
RelativeCostLogit_calcAverageCost
}
\details{
Calculate node profit rate using information from the
         normalization of children's shares. Note: we are using
         the lingo in C++ GCAM which is not intuitive.
}
\author{
KVC October 2017
}
