% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcImpliedCost}
\alias{RelativeCostLogit_calcImpliedCost}
\title{RelativeCostLogit_calcImpliedCost}
\usage{
RelativeCostLogit_calcImpliedCost(aShare, aCost, aPeriod)
}
\arguments{
\item{aShare}{TODO}

\item{aCost}{TODO}

\item{aPeriod}{Model time period}
}
\value{
Node profit rate (called IMPLIED_COST in this method)
}
\description{
RelativeCostLogit_calcImpliedCost
}
\details{
Calculate node profit rate (not sure why this is called "implied cost")
}
\author{
KVC September 2017
}
