% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcShareWeight}
\alias{RelativeCostLogit_calcShareWeight}
\title{RelativeCostLogit_calcShareWeight}
\usage{
RelativeCostLogit_calcShareWeight(aChoiceFnAbove, aShare, aProfit, aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{Choice function (logit type and exponent from node above)}

\item{aShare}{Share leaf has of node's land}

\item{aProfit}{Observed profit rate}

\item{aPeriod}{Model time period}
}
\value{
Share weight
}
\description{
RelativeCostLogit_calcShareWeight
}
\details{
Calculate the share weight using the relative cost logit
}
\author{
KVC September 2017
}
