% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcUnnormalizedShare}
\alias{RelativeCostLogit_calcUnnormalizedShare}
\title{RelativeCostLogit_calcUnnormalizedShare}
\usage{
RelativeCostLogit_calcUnnormalizedShare(aChoiceFnAbove, aShareWeight, aProfit,
  aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{Choice function (logit type and exponent from node above)}

\item{aShareWeight}{share weight for the choice for which the share is being calculated.}

\item{aProfit}{Profit for the choice for which the share is being calculated.}

\item{aPeriod}{model time period for the calculation.}
}
\value{
log of the unnormalized share.
}
\description{
RelativeCostLogit_calcUnnormalizedShare
}
\details{
Calculate the  numerator of the discrete choice (i.e., the unnormalized version)
         function being used to calculate child shares in this node. The normalization
         factor will be calculated later.
         Note: Negative costs can not be used in this logit formulation. Instead,
         the profit is capped at RelativeCostLogit::getMinCostThreshold.  This implies
         no behavior once profits have crossed this threshold.
         Note: this is modified to match the equations in GCAM4.3 (i.e., no log) but uses structure
         from versions of GCAM with the absolute cost logit.
}
\author{
KVC September 2017
}
