% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_getMinCostThreshold}
\alias{RelativeCostLogit_getMinCostThreshold}
\title{RelativeCostLogit_getMinCostThreshold}
\usage{
RelativeCostLogit_getMinCostThreshold()
}
\value{
The threshold value.
}
\description{
RelativeCostLogit_getMinCostThreshold
}
\details{
Get the minimum cost threshold value that may be used in this logit share equation.
}
\author{
KVC September 2017
}
