% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{run_bayes}
\alias{run_bayes}
\title{Run all the Bayesian analysis on a list of scenarios}
\usage{
run_bayes(aScenarioList)
}
\arguments{
\item{aScenarioList}{List of ScenarioInfo structures}
}
\value{
Modified list of ScenarioInfo structures with the Bayesian
calculation tables populated.
}
\description{
The assumption is that the scenarios have already been run.  The best way to
arrange this is to use the return value of \code{\link{run_ensemble}} as the
argument to this function
}
\section{TODO}{


\itemize{
\item{Offer some more control over likelihood and prior functions, xi values,
etc.}
\item{Offer control over years, land types.}
\item{Compute WAIC values in this function.}
}
}

