% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run an ensemble of offline land models}
\usage{
run_ensemble(N = 500, aOutputDir = "./outputs", skip = 0,
  atype = "Hindcast", logparallel = NULL)
}
\arguments{
\item{N}{Number of parameter sets to select}

\item{aOutputDir}{Output directory}

\item{skip}{Number of iterations to skip (i.e., if building on another run.)}

\item{atype}{Scenario type: either "Reference" or "Hindcast"}

\item{logparallel}{Name of directory to use for parallel workers' log files.
If \code{NULL}, then don't write log files.}
}
\value{
List of ScenarioInfo objects for the ensemble members
}
\description{
Parameter combinations are selected by generating a quasi-random
sequence and mapping it to a specified range for each parameter.
Then, each parameter set is run through the offline land model in
each of the Perfect, Lagged, and Linear variants.  (I.e., if N
parameter sets are selected, then 3N scenarios are run.)
}
\section{Output}{

The scenario results are written to a series of files in the specified output
directory.  There is a subdirectory for each output type, and output from
each scenario is written in its own file in the relevant subdirectory.  The
list of \code{ScenarioInfo} objects is written to a file called
\code{scenario-info.rds} in the output directory.  This file can be loaded
with a command such as \code{scenaro_list <-
readRDS('output/scenario-info.rds')}
}

\author{
KVC November 2017
}
