% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_model}
\alias{run_model}
\title{Run the GCAM land model}
\usage{
run_model(aScenarioInfo, aPeriods = NULL)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations.}

\item{aPeriods}{Integer vector of periods to run.  Default is all periods
defined for the scenario type.}
}
\value{
Name of the output directory for the run.
}
\description{
Loop through all years and run the land model.
}
\author{
KVC
}
