% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectivefun.R
\name{run_objective}
\alias{run_objective}
\title{Run user-specified objective function evaluations on a list of scenarios}
\usage{
run_objective(aScenarioList, years = NULL, landtypes = NULL,
  objectivefuns = c())
}
\arguments{
\item{aScenarioList}{List of ScenarioInfo structures}

\item{years}{Vector of years to filter observations to (default is to use all
available years)}

\item{landtypes}{Vector of land types to filter observations to (default is
to use all available land types)}
}
\value{
Modified list of ScenarioInfo structures with the objective
function calculation tables populated.
}
\description{
The assumption is that the scenarios have already been run.  The best way to
arrange this is to use the return value of \code{\link{run_ensemble}} as the
argument to this function
}
\details{
The \code{years} and \code{landtypes} arguments can be use to restrict the
observations that will be used in the analysis.  The regions are always
filtered to excatly the regions that are included in the ScenarioInfo
structures.

Typically log-probaility density functions are parameterized with a parameter
}
