% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\docType{data}
\name{SCENARIO.INFO}
\alias{SCENARIO.INFO}
\title{SCENARIO.INFO}
\format{
An object of class \code{ScenarioInfo} (inherits from \code{environment}) of length 26.
}
\usage{
SCENARIO.INFO
}
\description{
A \code{ScenarioInfo} object with parameters for the default scenario.
}
\author{
Kate Calvin
}
\keyword{datasets}
