% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{Sector_initCalc}
\alias{Sector_initCalc}
\title{Sector_initCalc}
\usage{
Sector_initCalc(
  aLandAllocator,
  aPeriod,
  aScenarioInfo,
  aLandConstraintCost = 0,
  aLandConstraintString = ""
)
}
\arguments{
\item{aLandAllocator}{Land allocator to perform initializations on}

\item{aPeriod}{Current model time period}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{aLandConstraintCost}{Additional cost used to constrain land}

\item{aLandConstraintString}{String identifying which leafs are included in the constraint}
}
\description{
Sector_initCalc
}
\details{
Initializations related to the sector
         Note: we are storing technology parameters in the Leafs for
         convenience, but are leaving the technology/leaf calculations
         separate as they are in the C++ code.
}
\author{
KVC September 2017
}
