% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_utils.R
\name{SectorUtils_normalizeLogShares}
\alias{SectorUtils_normalizeLogShares}
\title{SectorUtils_normalizeLogShares}
\usage{
SectorUtils_normalizeLogShares(aShares)
}
\arguments{
\item{aShares}{Unnormalized shares}
}
\value{
Normalized shares
}
\description{
SectorUtils_normalizeLogShares
}
\details{
Normalize log shares so they add up to one.
         Also, calculate parameters used to compute node profit.
}
\author{
KVC September 2017
}
