% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_utils.R
\name{SectorUtils_normalizeShares}
\alias{SectorUtils_normalizeShares}
\title{SectorUtils_normalizeShares}
\usage{
SectorUtils_normalizeShares(aShares)
}
\arguments{
\item{aShares}{Unnormalized shares}
}
\value{
Normalized shares and parameters needed to calculate node profit
}
\description{
SectorUtils_normalizeShares
}
\details{
Normalize shares so they add up to one.
}
\author{
KVC September 2017
}
