% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{softmax}
\alias{softmax}
\title{Compute the softmax function of a vector}
\usage{
softmax(x)
}
\arguments{
\item{x}{Vector of values for which to compute softmax}
}
\value{
Vector of softmax values
}
\description{
The softmax of a vector of variables \eqn{x_i} is
\deqn{s_i = \frac{\exp(x_i)}{\sum_{j=1}^{N} \exp(x_j)}}
}
