% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\docType{data}
\name{SRB.SCENARIO.INFO}
\alias{SRB.SCENARIO.INFO}
\title{SRB.SCENARIO.INFO}
\format{
An object of class \code{ScenarioInfo} (inherits from \code{environment}) of length 27.
}
\usage{
SRB.SCENARIO.INFO
}
\description{
A \code{ScenarioInfo} object with parameters for the PCHES scenario.
}
\author{
Kate Calvin
}
\keyword{datasets}
