% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{Subsector_initCalc}
\alias{Subsector_initCalc}
\title{Subsector_initCalc}
\usage{
Subsector_initCalc(aLandNode, aPeriod)
}
\arguments{
\item{aLandNode}{Land node to perform initializations on}

\item{aPeriod}{Current model time period}
}
\description{
Subsector_initCalc
}
\details{
Initializations related to the subsector.
         Note: we are using this to create a flexible nesting.
         This essentially serves as the "Node" level.
         This is a departure from how the C++ code works.
}
\author{
KVC September 2017
}
