% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sumx}
\alias{sumx}
\title{Sum values, minimizing roundoff error}
\usage{
sumx(x)
}
\arguments{
\item{x}{Values to be summed.}
}
\description{
This version of \code{sum} should be used when the values to be summed are
likely to differ dramatically in magnitude, such as when summing likelihood
values.
}
\keyword{internal}
