% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmanaged_land_leaf.R
\name{UnmanagedLandLeaf}
\alias{UnmanagedLandLeaf}
\title{UnmanagedLandLeaf}
\usage{
UnmanagedLandLeaf(aName, aFinalCalPeriod)
}
\arguments{
\item{aName}{Leaf name}

\item{aFinalCalPeriod}{Final calibration period}
}
\value{
New, initialized UnmanagedLandLeaf
}
\description{
UnmanagedLandLeaf
}
\details{
Initialize an Class called UnmanagedLandLeaf
}
\section{Fields}{

\describe{
\item{\code{mName}}{Leaf name}

\item{\code{mFinalCalPeriod}}{Final calibration period}

\item{\code{mLandAllocation}}{Land allocation for this leaf}

\item{\code{mShare}}{Share of land allocated to this leaf}

\item{\code{mShareWeight}}{Share weight of this leaf}

\item{\code{mProfitRate}}{Profit rate of this leaf}

\item{\code{mIsNewTech}}{Flag indicating leaf is a new technology (no calibration data)}
}}

\author{
KVC October 2017
}
