% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmanaged_land_leaf.R
\name{UnmanagedLandLeaf_initCalc}
\alias{UnmanagedLandLeaf_initCalc}
\title{UnmanagedLandLeaf_initCalc}
\usage{
UnmanagedLandLeaf_initCalc(aUnmanagedLandLeaf, aPeriod)
}
\arguments{
\item{aUnmanagedLandLeaf}{Unmanaged land leaf}

\item{aPeriod}{Current time period}
}
\description{
UnmanagedLandLeaf_initCalc
}
\details{
Initial calculations needed for the unmanagedland leaf.
}
\author{
KVC September 2017
}
