% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmanaged_land_leaf.R
\name{UnmanagedLandLeaf_setUnmanagedLandProfitRate}
\alias{UnmanagedLandLeaf_setUnmanagedLandProfitRate}
\title{UnmanagedLandLeaf_setUnmanagedLandProfitRate}
\usage{
UnmanagedLandLeaf_setUnmanagedLandProfitRate(
  aUnmanagedLandLeaf,
  aAverageProfitRate,
  aPeriod
)
}
\arguments{
\item{aUnmanagedLandLeaf}{Leaf to set profit rate for}

\item{aAverageProfitRate}{Average profit rate}

\item{aPeriod}{Time period}
}
\description{
UnmanagedLandLeaf_setUnmanagedLandProfitRate
}
\details{
This method adjusts the profit rate of an unmanaged land leaf
         to account for the carbon value of land if the ag subsidy is
         is active and a carbon price exists.
}
\author{
KVC October 2017
}
