% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\docType{data}
\name{YIELD.RATIOS}
\alias{YIELD.RATIOS}
\title{Ratio of Yield to First Historical Year}
\format{
An object of class \code{list} of length 7.
}
\usage{
YIELD.RATIOS
}
\description{
Yields at geopolitical region level for all FAO years
}
\author{
Kate Calvin
}
\keyword{datasets}
