% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hindcast_data.R
\docType{data}
\name{YIELDS.HIST}
\alias{YIELDS.HIST}
\title{Historic FAO Yields}
\format{
An object of class \code{data.frame} with 17292 rows and 4 columns.
}
\usage{
YIELDS.HIST
}
\description{
Yields at geopolitical region level for all FAO years
}
\author{
Kate Calvin
}
\keyword{datasets}
