% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_production_technology.R
\name{AgProductionTechnology_initCalc}
\alias{AgProductionTechnology_initCalc}
\title{AgProductionTechnology_initCalc}
\usage{
AgProductionTechnology_initCalc(aRegionName, aPeriod)
}
\arguments{
\item{aRegionName}{Region name.}

\item{aPeriod}{Model time period.}
}
\description{
AgProductionTechnology_initCalc
}
\details{
Initial calculation for the AgProductionTechnology.
         Adjusts costs & yields for technical change, calculate
         profit rate and make it available for the land allocator.
         Note: this method must be called before LandAllocator initCalc
}
\author{
KVC September 2017
}
