% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_price_data.R
\name{get_prices}
\alias{get_prices}
\title{get_prices}
\usage{
get_prices(aScenType)
}
\arguments{
\item{aScenType}{Type of scenario to run, either "Reference" or "Hindcast".}
}
\value{
Tibble containing prices by commodity and year
}
\description{
get_prices
}
\details{
Read in prices for all periods and return them
}
\author{
KVC October 2017
}
