% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{get_scenario_land_data}
\alias{get_scenario_land_data}
\title{Load land use results for an already-run scenario}
\usage{
get_scenario_land_data(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{ScenarioInfo structure for the run.}
}
\value{
Table with region, commodity, year, and area.
}
\description{
Fetch land use results and aggregate to region, commodity, and year.
}
\section{TODO}{


Currently the land use results don't have a region column, so we can't
aggregate by region (we have to assume that the results are all for a single
region).  Once we add a region column to the model output, we will want to
add region to the grouping variables for this function.
}

