% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian.R
\name{get_scenario_land_data}
\alias{get_scenario_land_data}
\title{Load land use results for an already-run scenario}
\usage{
get_scenario_land_data(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{ScenarioInfo structure for the run.}
}
\value{
Table with region, commodity, year, and area.
}
\description{
Fetch land use results and aggregate to region, commodity, and year.
}
