% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_timestep}
\alias{get_timestep}
\title{get_timestep}
\usage{
get_timestep(aPer)
}
\arguments{
\item{aPer}{Period to get timestep for}
}
\value{
Years
}
\description{
get_timestep
}
\details{
Calculate the length of the timestep leading up to a particular period
}
\author{
KVC October 2017
}
