% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\name{is.ScenarioInfo}
\alias{is.ScenarioInfo}
\title{Test whether an object is a \code{ScenarioInfo} object}
\usage{
is.ScenarioInfo(object)
}
\description{
Test whether an object is a \code{ScenarioInfo} object
}
