% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator}
\alias{LandAllocator}
\title{LandAllocator}
\usage{
LandAllocator(aRegionName, aFinalCalPeriod)
}
\arguments{
\item{aRegionName}{Region name}

\item{aFinalCalPeriod}{Final calibration period}
}
\value{
New, initialized LandAllocator
}
\description{
LandAllocator
}
\details{
Initialize an Class called LandAllocator
}
\section{Fields}{

\describe{
\item{\code{mRegionName}}{Region name}

\item{\code{mFinalCalPeriod}}{Final calibration period}

\item{\code{mChoiceFunction}}{Logit type & exponent for the top level of the nest}

\item{\code{mLandAllocation}}{Land allocation for this region}

\item{\code{mShare}}{Share of land}

\item{\code{mUnmanagedLandValue}}{Unmanaged land value for the allocator}

\item{\code{mChildren}}{Children of the LandAllocator}
}}

\author{
KVC September 2017
}
