% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator}
\alias{LandAllocator}
\title{LandAllocator}
\usage{
LandAllocator(aRegionName, aLogitExponent, aLandAllocation)
}
\arguments{
\item{aRegionName}{Region name}

\item{aLogitExponent}{Logit exponent of the top level of the land allocator}

\item{aLandAllocation}{Land allocation for this region}
}
\value{
New, initialized LandAllocator
}
\description{
LandAllocator
}
\details{
Initialize an Class called LandAllocator
}
\section{Fields}{

\describe{
\item{\code{mRegionName}}{Region name}

\item{\code{mLogitExponent}}{Logit exponent for the top level of the nest}

\item{\code{mLandAllocation}}{Land allocation for this region}

\item{\code{mShare}}{Share of land}

\item{\code{mChild}}{Children of the LandAllocator (currently one LandNode only)}
}}

\author{
KVC September 2017
}
