% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_calcLandAllocation}
\alias{LandAllocator_calcLandAllocation}
\title{LandAllocator_calcLandAllocation}
\usage{
LandAllocator_calcLandAllocation(aLandAllocator, aPeriod)
}
\arguments{
\item{aLandAllocator}{Land allocator for this region}

\item{aPeriod}{Model time period.}
}
\description{
LandAllocator_calcLandAllocation
}
\details{
Initiates calculation of land allocation by
         calling calcLandAllocation on all children
}
\author{
KVC September 2017
}
