% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getExpectedYield}
\alias{LandAllocator_getExpectedYield}
\title{LandAllocator_getExpectedYield}
\usage{
LandAllocator_getExpectedYield(aLandAllocator, aData, scentype)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aData}{Data table to store expected yield}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Expected yield data
}
\description{
LandAllocator_getExpectedYield
}
\details{
Calculates and returns expected yield for all leafs
}
\author{
KVC October 2017
}
