% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getLandAllocation}
\alias{LandAllocator_getLandAllocation}
\title{LandAllocator_getLandAllocation}
\usage{
LandAllocator_getLandAllocation(aLandAllocator, allLand, scentype)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{allLand}{Data frame to fill in land allocation}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Land allocation table
}
\description{
LandAllocator_getLandAllocation
}
\details{
Calculates and returns land allocation for a particular leaf
}
\author{
KVC October 2017
}
