% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_getLandAllocation}
\alias{LandAllocator_getLandAllocation}
\title{LandAllocator_getLandAllocation}
\usage{
LandAllocator_getLandAllocation(aLandAllocator, aName, aPeriod)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aName}{Name of leaf we want allocation for}

\item{aPeriod}{Model period}
}
\value{
Land allocation
}
\description{
LandAllocator_getLandAllocation
}
\details{
Calculates and returns land allocation for a particular leaf
}
\author{
KVC October 2017
}
