% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandAllocator_getYield}
\alias{LandAllocator_getYield}
\title{LandAllocator_getYield}
\usage{
LandAllocator_getYield(aLandAllocator, aName, aPeriod)
}
\arguments{
\item{aLandAllocator}{LandAllocator}

\item{aName}{Name of leaf we want yield for}

\item{aPeriod}{Model period}
}
\value{
Yield
}
\description{
LandAllocator_getYield
}
\details{
Calculates and returns yield for a particular leaf
}
\author{
KVC October 2017
}
