% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_initCalc}
\alias{LandAllocator_initCalc}
\title{LandAllocator_initCalc}
\usage{
LandAllocator_initCalc(aLandAllocator, aPeriod)
}
\arguments{
\item{aLandAllocator}{The land allocator for this region}

\item{aPeriod}{Current time period}
}
\description{
LandAllocator_initCalc
}
\details{
Initial calculations needed for the land allocator.
}
\author{
KVC September 2017
}
