% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_allocator.R
\name{LandAllocator_setInitShares}
\alias{LandAllocator_setInitShares}
\title{LandAllocator_setInitShares}
\usage{
LandAllocator_setInitShares(aRegionName, aLandAllocationAbove, aPeriod)
}
\arguments{
\item{aRegionName}{Region name.}

\item{aLandAllocationAbove}{Land allocation of the parent node.}

\item{aPeriod}{model period.}
}
\description{
LandAllocator_setInitShares
}
\details{
Calculates and sets initial land shares
}
\author{
KVC September 2017
}
