% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LandAllocator_setup}
\alias{LandAllocator_setup}
\title{LandAllocator_setup}
\usage{
LandAllocator_setup(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}
}
\description{
LandAllocator_setup
}
\details{
Setup the land allocator. This includes
         Reading in external data and organizing
         it in the node, leaf, technology structure
}
\author{
KVC October 2017
}
