% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf}
\alias{LandLeaf}
\title{LandLeaf}
\usage{
LandLeaf(aName, aFinalCalPeriod)
}
\arguments{
\item{aName}{Leaf name}

\item{aFinalCalPeriod}{Final calibration period}
}
\value{
New, initialized LandLeaf
}
\description{
LandLeaf
}
\details{
Initialize an Class called LandLeaf
}
\section{Fields}{

\describe{
\item{\code{mName}}{Leaf name}

\item{\code{mFinalCalPeriod}}{Final calibration period}

\item{\code{mProductName}}{Name of product produced by this leaf}

\item{\code{mLandAllocation}}{Land allocation for this leaf}

\item{\code{mCalLandAllocation}}{Calibration land allocation for this leaf}

\item{\code{mShare}}{Share of land allocated to this leaf}

\item{\code{mShareWeight}}{Share weight of this leaf}

\item{\code{mProfitRate}}{Profit rate of this leaf}

\item{\code{mCost}}{Non-land variable cost of this leaf}

\item{\code{mYield}}{yield for this leaf}

\item{\code{mExpectedYield}}{expected yield for this leaf}

\item{\code{mExpectedPrice}}{expected price for this leaf}

\item{\code{mCalOutput}}{calibrated-output for this leaf}

\item{\code{mNonLandCostTechChange}}{Technical change on the cost of this leaf}

\item{\code{mAgProdChange}}{Technical change on yield for this leaf}
}}

\author{
KVC September 2017
}
