% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf}
\alias{LandLeaf}
\title{LandLeaf}
\usage{
LandLeaf(aName, aLandAllocation)
}
\arguments{
\item{aName}{Leaf name}

\item{aLandAllocation}{Land allocation for this leaf}
}
\value{
New, initialized LandLeaf
}
\description{
LandLeaf
}
\details{
Initialize an Class called LandLeaf
}
\section{Fields}{

\describe{
\item{\code{mName}}{Leaf name}

\item{\code{mLandAllocation}}{Land allocation for this leaf}

\item{\code{mShare}}{Share of land allocated to this leaf}

\item{\code{mShareWeight}}{Share weight of this leaf}

\item{\code{mProfitRate}}{Profit rate of this leaf}
}}

\author{
KVC September 2017
}
