% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_calculateShareWeight}
\alias{LandLeaf_calculateShareWeight}
\title{LandLeaf_calculateShareWeight}
\usage{
LandLeaf_calculateShareWeight(aLandLeaf, aChoiceFnAbove, aPeriod, NODE_PROFIT)
}
\arguments{
\item{aLandLeaf}{Land leaf to perform calculations on}

\item{aChoiceFnAbove}{Type of logit}

\item{aPeriod}{Model time period}

\item{NODE_PROFIT}{Profit rate of node (TODO: put this in the choice function)}
}
\description{
LandLeaf_calculateShareWeight
}
\details{
Calculates share weights for land leafs
}
\author{
KVC September 2017
}
