% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_calculateShareWeight}
\alias{LandLeaf_calculateShareWeight}
\title{LandLeaf_calculateShareWeight}
\usage{
LandLeaf_calculateShareWeight(aLandLeaf, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{Type of logit}

\item{aPeriod}{Model time period}

\item{aRegionName}{Region name}
}
\description{
LandLeaf_calculateShareWeight
}
\details{
Calculates share weights for land leafs
}
\author{
KVC September 2017
}
