% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandLeaf_getLandAllocation}
\alias{LandLeaf_getLandAllocation}
\title{LandLeaf_getLandAllocation}
\usage{
LandLeaf_getLandAllocation(aLandLeaf, allLand, scentype)
}
\arguments{
\item{aLandLeaf}{LandLeaf}

\item{allLand}{Data frame to fill in land allocation}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Land allocation table
}
\description{
LandLeaf_getLandAllocation
}
\details{
Calculates and returns total land allocation for types and periods
}
\author{
KVC October 2017
}
