% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_leaf.R
\name{LandLeaf_initCalc}
\alias{LandLeaf_initCalc}
\title{LandLeaf_initCalc}
\usage{
LandLeaf_initCalc(aLandLeaf, aPeriod)
}
\arguments{
\item{aLandLeaf}{Land leaf}

\item{aPeriod}{Current time period}
}
\description{
LandLeaf_initCalc
}
\details{
Initial calculations needed for the land leaf.
}
\author{
KVC September 2017
}
