% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode}
\alias{LandNode}
\title{LandNode}
\usage{
LandNode(aName, aChoiceFunction, aLandAllocation, aFinalCalPeriod)
}
\arguments{
\item{aName}{Node name}

\item{aChoiceFunction}{Choice function (logit type and exponent from node above)}

\item{aLandAllocation}{Land allocation for this node (reserved for future
use)}

\item{aFinalCalPeriod}{Final calibration period}
}
\value{
New, initialized LandNode
}
\description{
LandNode
}
\details{
Initialize an Class called LandNode
}
\section{Fields}{

\describe{
\item{\code{mName}}{Node name}

\item{\code{mChoiceFunction}}{Choice function (logit type and exponent) for this node}

\item{\code{mLandAllocation}}{Land allocation for this node}

\item{\code{mFinalCalPeriod}}{Final calibration period}

\item{\code{mUnmanagedLandValue}}{Unmanaged land value in this node}

\item{\code{mShare}}{Share of land allocated to this node}

\item{\code{mShareWeight}}{Share weight of this node}

\item{\code{mGhostUnnormalizedShare}}{Ghost share for new children in this node}

\item{\code{mProfitRate}}{Profit rate of this node}

\item{\code{mChildren}}{list of LandLeaf children}
}}

\author{
KVC September 2017
}
