% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode}
\alias{LandNode}
\title{LandNode}
\usage{
LandNode(aName, aChoiceFunction, aLandAllocation)
}
\arguments{
\item{aName}{Node name}

\item{aLandAllocation}{Land allocation for this node}

\item{aLogitExponent}{Logit exponent of the node}
}
\value{
New, initialized LandNode
}
\description{
LandNode
}
\details{
Initialize an Class called LandNode
}
\section{Fields}{

\describe{
\item{\code{mName}}{Node name}

\item{\code{mLogitExponent}}{Logit exponent of the node}

\item{\code{mLandAllocation}}{Land allocation for this node}

\item{\code{mShare}}{Share of land allocated to this node}

\item{\code{mShareWeight}}{Share weight of this node}

\item{\code{mProfitRate}}{Profit rate of this node}

\item{\code{mChildren}}{list of LandLeaf children}
}}

\author{
KVC September 2017
}
