% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_calculateShareWeight}
\alias{LandNode_calculateShareWeight}
\title{LandNode_calculateShareWeight}
\usage{
LandNode_calculateShareWeight(aLandNode, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aLandNode}{Land node}

\item{aChoiceFnAbove}{Type of logit}

\item{aPeriod}{Model time period}
}
\description{
LandNode_calculateShareWeight
}
\details{
Calculates share weights for land nodes
}
\author{
KVC September 2017
}
