% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_calculateShareWeights}
\alias{LandNode_calculateShareWeights}
\title{LandNode_calculateShareWeights}
\usage{
LandNode_calculateShareWeights(aLandNode, aChoiceFnAbove, aPeriod)
}
\arguments{
\item{aLandNode}{Land node to perform calculations on}

\item{aChoiceFnAbove}{The discrete choice function from the level above.}

\item{aPeriod}{Period.}
}
\description{
LandNode_calculateShareWeights
}
\details{
Calculates share weights at the node level and calls
         the share weight calculation for the node's children.
         Note: this now blends the calculateCalibrationProfitRate and
         calculateProfitScaler code in GCAM4.3. It will be helpful when
         implementing an absolute cost logit.
}
\author{
KVC September 2017
}
