% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getExpectedYield}
\alias{LandNode_getExpectedYield}
\title{LandNode_getExpectedYield}
\usage{
LandNode_getExpectedYield(aLandNode, aData, scentype)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aData}{Data table to store expected yield}

\item{scentype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
Expected yield data
}
\description{
LandNode_getExpectedYield
}
\details{
Calculates and returns expected yield for leafs in this node
}
\author{
KVC October 2017
}
