% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_node.R
\name{LandNode_getLandAllocation}
\alias{LandNode_getLandAllocation}
\title{LandNode_getLandAllocation}
\usage{
LandNode_getLandAllocation(aLandNode, aName, aPeriod)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aPeriod}{Model period}
}
\value{
Land allocation for this node
}
\description{
LandNode_getLandAllocation
}
\details{
Calculates and returns total land allocation of a given type.
}
\author{
KVC October 2017
}
