% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getLandShares}
\alias{LandNode_getLandShares}
\title{LandNode_getLandShares}
\usage{
LandNode_getLandShares(aLandNode, aShares, aPeriod)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aShares}{Table of shares to add to}

\item{aPeriod}{Model period}
}
\value{
Table of land shares
}
\description{
LandNode_getLandShares
}
\details{
Calculates and returns land share for all children.
}
\author{
KVC November 2017
}
