% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{LandNode_getYield}
\alias{LandNode_getYield}
\title{LandNode_getYield}
\usage{
LandNode_getYield(aLandNode, aName, aPeriod)
}
\arguments{
\item{aLandNode}{LandNode}

\item{aName}{Name of leaf we want yield for}

\item{aPeriod}{Model period}
}
\value{
Yield for leaf named `aName`
}
\description{
LandNode_getYield
}
\details{
Calculates and returns yield for a given leaf
}
\author{
KVC October 2017
}
