% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LandNode_setup}
\alias{LandNode_setup}
\title{LandNode_setup}
\usage{
LandNode_setup(aLandAllocator, aRegionName, aData, aColumnName)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}

\item{aRegionName}{Region}

\item{aData}{Data needed for setting up this node}

\item{aColumnName}{Column name with the parent}
}
\description{
LandNode_setup
}
\details{
Setup a LandNode level of the land allocator. This includes
         reading in external data, creating LandNode nodes.
}
\author{
KVC October 2017
}
