% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{Leaf_setup}
\alias{Leaf_setup}
\title{Leaf_setup}
\usage{
Leaf_setup(aLandAllocator, aRegionName, data, col.name, ag.data = NULL)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}

\item{aRegionName}{Region}

\item{data}{Data needed for setting up this node}

\item{col.name}{Column name with the parent}

\item{ag.data}{Agricultural technology data}
}
\description{
Leaf_setup
}
\details{
Setup land leafs for the land allocator. This includes
         searching for the right parent, creating and creating leafs.
}
\author{
KVC October 2017
}
