% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{Leaf_setup}
\alias{Leaf_setup}
\title{Leaf_setup}
\usage{
Leaf_setup(aLandAllocator, aRegionName, aData, aColName, aScenarioInfo,
  aAgData = NULL, newTechData = NULL)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}

\item{aRegionName}{Region}

\item{aData}{Data needed for setting up this node}

\item{aColName}{Column name with the parent}

\item{aScenarioInfo}{ScenarioInfo structure containing the scenario parameters}

\item{aAgData}{Agricultural technology data}

\item{newTechData}{Data about any new technologies (i.e., technologies without calibration information)}
}
\description{
Leaf_setup
}
\details{
Setup land leafs for the land allocator. This includes
         searching for the right parent, creating and creating leafs.
}
\author{
KVC October 2017
}
