% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{LN1_setup}
\alias{LN1_setup}
\title{LN1_setup}
\usage{
LN1_setup(aLandAllocator, aRegionName, data, col.name)
}
\arguments{
\item{aLandAllocator}{LandAllocator that needs set up}
}
\description{
LN1_setup
}
\details{
Setup LN1 of the land allocator. This includes
         reading in external data, creating LandNode1 nodes, and
         creating any leafs with only one parent node.
}
\author{
KVC October 2017
}
