% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plotLandAllocation}
\alias{plotLandAllocation}
\title{plotLandAllocation}
\usage{
plotLandAllocation(aScenarioInfo)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
plotLandAllocation
}
\details{
Plots land allocation over time
}
\author{
KVC September 2017
}
