% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land.R
\name{plotRegionalLandAllocation}
\alias{plotRegionalLandAllocation}
\title{plotRegionalLandAllocation}
\usage{
plotRegionalLandAllocation(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land Allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
plotRegionalLandAllocation
}
\details{
Plots regional land allocation over time.
         Aggregates land by region (i.e., adds all AEZs together)
         and then plots land.
}
\author{
KVC October 2017
}
