% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_price_data.R
\docType{data}
\name{PRICES}
\alias{PRICES}
\title{Price tables for each scenario type.}
\format{An object of class \code{list} of length 2.}
\usage{
PRICES
}
\description{
Currently supported types are "Reference" and "Hindcast".  This structure is
a list of tibbles with all of the prices for the model, for each scenario.
}
\author{
Kate Calvin, Robert Link
}
\keyword{datasets}
