% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printAllOutputs}
\alias{printAllOutputs}
\title{printAllOutputs}
\usage{
printAllOutputs(aLandAllocator, aScenarioInfo, aNest,
  aFileOutput = FALSE)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{aNest}{Nest to fill in}

\item{aFileOutput}{Flag indicating whether output should be written to a
file, in addition to being returned.}
}
\value{
Table of all model outputs.
}
\description{
printAllOutputs
}
\details{
Prints all outputs (land allocation, expected price, expected yield, yield) by land leaf
}
\author{
KVC May 2018
}
