% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printExpectedPrices}
\alias{printExpectedPrices}
\title{printExpectedPrices}
\usage{
printExpectedPrices(aScenarioInfo)
}
\description{
printExpectedPrices
}
\details{
Print expected prices
}
\author{
KVC November 2017
}
