% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printLandShares}
\alias{printLandShares}
\title{printLandShares}
\usage{
printLandShares(aLandAllocator, aScenarioInfo, aNest)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{aNest}{Current land nest}
}
\description{
printLandShares
}
\details{
Prints land share for all nodes and leafs
}
\author{
KVC November 2017
}
