% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printNest}
\alias{printNest}
\title{printNest}
\usage{
printNest(aLandAllocator, aScenarioInfo, aFilter = "Urban|Rock|Tundra")
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}

\item{aFilter}{Regular expression to filter out of graph output (the return
value contains the full table, regardless of filters)}
}
\value{
Table of land node nestings, constructed by \code{\link{getNest}}.
}
\description{
printNest
}
\details{
Prints the land allocator nest, which can
         be used to plot tree structure or to query
         leaf data
}
\author{
KVC October 2017
}
