% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{printOutput}
\alias{printOutput}
\title{printLandAllocation}
\usage{
printOutput(aLandAllocator, aScenarioInfo)
}
\arguments{
\item{aLandAllocator}{Land allocator}

\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\description{
printLandAllocation
}
\details{
Prints all outputs
}
\author{
KVC October 2017
}
