% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{pwaic}
\alias{pwaic}
\title{Compute the effective number of parameters \code{pwaic} for a model.}
\usage{
pwaic(pointwise_ll)
}
\arguments{
\item{pointwise_ll}{List returned from \code{\link{pointwise_loglike}}.}
}
\value{
A vector of pwaic values averaged over Monte Carlo samples.  Each
observation will have one corresponding point in the vector.
}
\description{
For each observation, compute the variance of the log likelihood.  Note that
we are computing the variance of the \emph{log} likelihood.  The total
\code{pwaic} for the model is the sum of these, but we return the individual
values for use in computing standard errors.
}
\keyword{internal}
