% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN1_Node}
\alias{ReadData_LN1_Node}
\title{ReadData_LN1_Node}
\usage{
ReadData_LN1_Node(aRegionName, SUBREGION = NULL, subregionData = NULL)
}
\arguments{
\item{aRegionName}{Region to read data for}

\item{SUBREGION}{Subregion to read data for}

\item{subregionData}{Subregion data}
}
\value{
Data on children of the land allocator
}
\description{
ReadData_LN1_Node
}
\details{
Read in unmanaged land value, names of children, and logit exponents
         for the LandAllocator.
}
\author{
KVC October 2017
}
