% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData_LN2_Node}
\alias{ReadData_LN2_Node}
\title{ReadData_LN2_Node}
\usage{
ReadData_LN2_Node(aRegionName, aSubRegion, subregionData)
}
\arguments{
\item{aRegionName}{Region to read data for}

\item{aSubRegion}{Subregion to read data for}

\item{subregionData}{Subregion data}
}
\value{
Data on level 2 nodes of the land allocator
}
\description{
ReadData_LN2_Node
}
\details{
Read in names of children and logit exponents for LN2
}
\author{
KVC October 2017
}
