% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcImpliedCost}
\alias{RelativeCostLogit_calcImpliedCost}
\title{RelativeCostLogit_calcImpliedCost}
\usage{
RelativeCostLogit_calcImpliedCost(aChoiceFnAbove, aShare, aCost, aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{Choice function (logit type and exponent from node above)}

\item{aShare}{share of node}

\item{aCost}{default profit of node}

\item{aPeriod}{Model time period}
}
\value{
Node profit rate (called `impliedCost` in this method)
}
\description{
RelativeCostLogit_calcImpliedCost
}
\details{
Calculate node profit rate (not sure why this is called "implied cost")
}
\author{
KVC September 2017
}
