% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcShareWeight}
\alias{RelativeCostLogit_calcShareWeight}
\title{RelativeCostLogit_calcShareWeight}
\usage{
RelativeCostLogit_calcShareWeight(aShare, aCost, aPeriod, OUTPUT_COST)
}
\arguments{
\item{aShare}{TODO}

\item{aCost}{TODO}

\item{aPeriod}{Model time period.}
}
\value{
Share weight
}
\description{
RelativeCostLogit_calcShareWeight
}
\details{
Calculate the share weight using the relative cost logit
}
\author{
KVC September 2017
}
