% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_cost_logit.R
\name{RelativeCostLogit_calcUnnormalizedShare}
\alias{RelativeCostLogit_calcUnnormalizedShare}
\title{RelativeCostLogit_calcUnnormalizedShare}
\usage{
RelativeCostLogit_calcUnnormalizedShare(aChoiceFnAbove, aShareWeight, aCost,
  aPeriod)
}
\arguments{
\item{aChoiceFnAbove}{Choice function (logit type and exponent from node above)}

\item{aShareWeight}{share weight for the choice for which the share is being calculated.}

\item{aCost}{cost for the choice for which the share is being calculated.}

\item{aPeriod}{model time period for the calculation.}
}
\value{
log of the unnormalized share.
}
\description{
RelativeCostLogit_calcUnnormalizedShare
}
\details{
Calculate the log of the numerator of the discrete choice (i.e., the unnormalized version)
         function being used to calculate subsector shares in this sector.  The normalization
         factor will be calculated later.
         Note: Negative costs can not be used in this logit formulation.  Instead the cost
         the cost is capped at RelativeCostLogit::getMinCostThreshold.  This implies
         no behavior once costs have crossed this threshold.
}
\author{
KVC September 2017
}
