% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{data}
\name{REQD.SUBDIRS}
\alias{REQD.SUBDIRS}
\title{Required subdirectories for output directory}
\format{An object of class \code{NULL} of length 0.}
\usage{
REQD.SUBDIRS
}
\description{
Required subdirectories for output directory
}
\keyword{internal}
