% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_ensemble}
\alias{run_ensemble}
\title{Run an ensemble of offline land models}
\usage{
run_ensemble(N = 500, aOutputDir = "./outputs", atype = "Hindcast")
}
\arguments{
\item{N}{Number of parameter sets to select}

\item{aOutputDir}{Output directory}

\item{atype}{Scenario type: either "Reference" or "Hindcast"}
}
\value{
List of ScenarioInfo objects for the ensemble members
}
\description{
Parameter combinations are selected by generating a quasi-random
sequence and mapping it to a specified range for each parameter.
Then, each parameter set is run through the offline land model in
each of the Perfect, Lagged, and Linear variants.  (I.e., if N
parameter sets are selected, then 3N scenarios are run.)
}
\author{
KVC November 2017
}
