% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_ensembles}
\alias{run_ensembles}
\title{run_ensembles}
\usage{
run_ensembles(aOutputDir = "./outputs")
}
\arguments{
\item{aOutputDir}{Output directory}
}
\description{
run_ensembles
}
\details{
Loop over a large parameter set and run the offline land model
}
\author{
KVC November 2017
}
