% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(aScenarioInfo, aPeriods = PERIODS)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations}
}
\value{
Name of the output directory
}
\description{
run_model
}
\details{
Loops through all years and runs the land model.
}
\author{
KVC
}
