% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_model}
\alias{run_model}
\title{Run the GCAM land model}
\usage{
run_model(aScenarioInfo, aPeriods = NULL, aVerbose = FALSE)
}
\arguments{
\item{aScenarioInfo}{Scenario-related information, including names, logits, expectations.}

\item{aPeriods}{Integer vector of periods to run.  Default is all periods
defined for the scenario type.}

\item{aVerbose}{If \code{TRUE}, output additional debugging information.}
}
\value{
Table of model results.
}
\description{
Loop through all years and run the land model.
}
\examples{
run_model(SCENARIO.INFO, aVerbose=TRUE)
run_model(SCENARIO.INFO, aPeriods = 1:5)
}
\author{
KVC
}
