% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{Sector_initCalc}
\alias{Sector_initCalc}
\title{Sector_initCalc}
\usage{
Sector_initCalc(aLandAllocator, aPeriod)
}
\arguments{
\item{aLandAllocator}{Land allocator to perform initializations on}

\item{aPeriod}{Current model time period}
}
\description{
Sector_initCalc
}
\details{
Initializations related to the technology.
         Note: we are storing technology parameters in the Leafs for
         convenience, but are leaving the technology/leaf calculations
         separate as they are in the C++ code.
}
\author{
KVC September 2017
}
