% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\name{update_scen_info}
\alias{update_scen_info}
\title{update_scen_info}
\usage{
update_scen_info(aName = NULL, aScenarioType = DEFAULT.SCENARIO.TYPE,
  aExpectationType = "Perfect", aLinearYears = NULL,
  aLaggedShareOld = NULL, aUseZeroCost = FALSE)
}
\arguments{
\item{aName}{New scenario name (default will generate this from other info)}

\item{aScenarioType}{New scenario type (default = `DEFAULT.SCENARIO.TYPE`)}

\item{aExpectationType}{New expectation type (default = Perfect)}

\item{aLinearYears}{New linear years (default = NULL)}

\item{aLaggedShareOld}{New lagged share old (default = NULL)}

\item{aUseZeroCost}{New cost assumption (default = FALSE)}
}
\value{
Updated scenario info object
}
\description{
update_scen_info
}
\author{
KVC November 2018
}
