% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scenario_info.R
\name{update_scen_info}
\alias{update_scen_info}
\title{update_scen_info}
\usage{
update_scen_info(aName = NULL, aScenarioType = DEFAULT.SCENARIO.TYPE,
  aExpectationType = "Perfect", aLinearYears = NULL,
  aLaggedShareOld = NULL, aUseZeroCost = FALSE, aCalibrateShareWt = TRUE,
  aShareWts = NULL)
}
\arguments{
\item{aName}{New scenario name (default will generate this from other info)}

\item{aScenarioType}{New scenario type (default = `DEFAULT.SCENARIO.TYPE`)}

\item{aExpectationType}{New expectation type (default = Perfect)}

\item{aLinearYears}{New linear years (default = NULL)}

\item{aLaggedShareOld}{New lagged share old (default = NULL)}

\item{aUseZeroCost}{New cost assumption (default = FALSE)}

\item{aCalibrateShareWt}{Flag indicating share weights should be calibrated}

\item{aShareWts}{Vector of share weights}
}
\value{
Updated scenario info object
}
\description{
This function takes the default `SCENARIO.INFO` object and updates it based on
user specified arguments. Some updates (e.g., changing from "Reference" to "Hindcast" scenario
type) will automatically update the scenario name and file name.
}
\examples{
update_scen_info(aCalibrateShareWt = FALSE, aShareWts=get_saved_share_weights())
update_scen_info(aScenarioType = "Hindcast")
}
\author{
KVC November 2018
}
